linux_api

Module exit

Functions

exit

Exits the current thread, setting val as the exit code.

exit_group

Exits the current process, setting val as the exit code.

exit_group_raw

Exits the process, setting val & 0xff as the exit code.

exit_raw

Exits the current thread, setting val & 0xff as the exit code.