Enum linux_api::ioctls::IoctlRequest

#[repr(u32)]
pub enum IoctlRequest {

    TCGETS = 21_505,
    TCSETS = 21_506,
    TCSETSW = 21_507,
    TCSETSF = 21_508,
    TCGETA = 21_509,
    TCSETA = 21_510,
    TCSETAW = 21_511,
    TCSETAF = 21_512,
    TCSBRK = 21_513,
    TCXONC = 21_514,
    TCFLSH = 21_515,
    TIOCEXCL = 21_516,
    TIOCNXCL = 21_517,
    TIOCSCTTY = 21_518,
    TIOCGPGRP = 21_519,
    TIOCSPGRP = 21_520,
    TIOCOUTQ = 21_521,
    TIOCSTI = 21_522,
    TIOCGWINSZ = 21_523,
    TIOCSWINSZ = 21_524,
    TIOCMGET = 21_525,
    TIOCMBIS = 21_526,
    TIOCMBIC = 21_527,
    TIOCMSET = 21_528,
    TIOCGSOFTCAR = 21_529,
    TIOCSSOFTCAR = 21_530,
    FIONREAD = 21_531,
    TIOCLINUX = 21_532,
    TIOCCONS = 21_533,
    TIOCGSERIAL = 21_534,
    TIOCSSERIAL = 21_535,
    TIOCPKT = 21_536,
    FIONBIO = 21_537,
    TIOCNOTTY = 21_538,
    TIOCSETD = 21_539,
    TIOCGETD = 21_540,
    TCSBRKP = 21_541,
    TIOCSBRK = 21_543,
    TIOCCBRK = 21_544,
    TIOCGSID = 21_545,
    TIOCGRS485 = 21_550,
    TIOCSRS485 = 21_551,
    TCGETX = 21_554,
    TCSETX = 21_555,
    TCSETXF = 21_556,
    TCSETXW = 21_557,
    TIOCVHANGUP = 21_559,
    FIONCLEX = 21_584,
    FIOCLEX = 21_585,
    FIOASYNC = 21_586,
    TIOCSERCONFIG = 21_587,
    TIOCSERGWILD = 21_588,
    TIOCSERSWILD = 21_589,
    TIOCGLCKTRMIOS = 21_590,
    TIOCSLCKTRMIOS = 21_591,
    TIOCSERGSTRUCT = 21_592,
    TIOCSERGETLSR = 21_593,
    TIOCSERGETMULTI = 21_594,
    TIOCSERSETMULTI = 21_595,
    TIOCMIWAIT = 21_596,
    TIOCGICOUNT = 21_597,
    FIOQSIZE = 21_600,
    FIOSETOWN = 35_073,
    SIOCSPGRP = 35_074,
    FIOGETOWN = 35_075,
    SIOCGPGRP = 35_076,
    SIOCATMARK = 35_077,
    SOCK_IOC_TYPE = 137,
    EPOLL_IOC_TYPE = 138,
    SIOCGSTAMP = 35_078,
    SIOCGSTAMPNS = 35_079,
    SIOCADDRT = 35_083,
    SIOCDELRT = 35_084,
    SIOCRTMSG = 35_085,
    SIOCGIFNAME = 35_088,
    SIOCSIFLINK = 35_089,
    SIOCGIFCONF = 35_090,
    SIOCGIFFLAGS = 35_091,
    SIOCSIFFLAGS = 35_092,
    SIOCGIFADDR = 35_093,
    SIOCSIFADDR = 35_094,
    SIOCGIFDSTADDR = 35_095,
    SIOCSIFDSTADDR = 35_096,
    SIOCGIFBRDADDR = 35_097,
    SIOCSIFBRDADDR = 35_098,
    SIOCGIFNETMASK = 35_099,
    SIOCSIFNETMASK = 35_100,
    SIOCGIFMETRIC = 35_101,
    SIOCSIFMETRIC = 35_102,
    SIOCGIFMEM = 35_103,
    SIOCSIFMEM = 35_104,
    SIOCGIFMTU = 35_105,
    SIOCSIFMTU = 35_106,
    SIOCSIFNAME = 35_107,
    SIOCSIFHWADDR = 35_108,
    SIOCGIFENCAP = 35_109,
    SIOCSIFENCAP = 35_110,
    SIOCGIFHWADDR = 35_111,
    SIOCGIFSLAVE = 35_113,
    SIOCSIFSLAVE = 35_120,
    SIOCADDMULTI = 35_121,
    SIOCDELMULTI = 35_122,
    SIOCGIFINDEX = 35_123,
    SIOCSIFPFLAGS = 35_124,
    SIOCGIFPFLAGS = 35_125,
    SIOCDIFADDR = 35_126,
    SIOCSIFHWBROADCAST = 35_127,
    SIOCGIFCOUNT = 35_128,
    SIOCGIFBR = 35_136,
    SIOCSIFBR = 35_137,
    SIOCGIFTXQLEN = 35_138,
    SIOCSIFTXQLEN = 35_139,
    SIOCETHTOOL = 35_142,
    SIOCGMIIPHY = 35_143,
    SIOCGMIIREG = 35_144,
    SIOCSMIIREG = 35_145,
    SIOCWANDEV = 35_146,
    SIOCOUTQNSD = 35_147,
    SIOCGSKNS = 35_148,
    SIOCDARP = 35_155,
    SIOCGARP = 35_156,
    SIOCSARP = 35_157,
    SIOCDRARP = 35_168,
    SIOCGRARP = 35_169,
    SIOCSRARP = 35_170,
    SIOCGIFMAP = 35_184,
    SIOCSIFMAP = 35_185,
    SIOCADDDLCI = 35_200,
    SIOCDELDLCI = 35_201,
    SIOCGIFVLAN = 35_202,
    SIOCSIFVLAN = 35_203,
    SIOCBONDENSLAVE = 35_216,
    SIOCBONDRELEASE = 35_217,
    SIOCBONDSETHWADDR = 35_218,
    SIOCBONDSLAVEINFOQUERY = 35_219,
    SIOCBONDINFOQUERY = 35_220,
    SIOCBONDCHANGEACTIVE = 35_221,
    SIOCBRADDBR = 35_232,
    SIOCBRDELBR = 35_233,
    SIOCBRADDIF = 35_234,
    SIOCBRDELIF = 35_235,
    SIOCSHWTSTAMP = 35_248,
    SIOCGHWTSTAMP = 35_249,
    SIOCDEVPRIVATE = 35_312,
    SIOCPROTOPRIVATE = 35_296,

}

ioctl requests

While the ioctl syscall takes the request as an unsigned long, which is 64-bits on x86-64, the constants are specifically documented as being 32-bits, so we use an underlying representation of 32-bits.

Variants

TCGETS = 21_505

TCSETS = 21_506

TCSETSW = 21_507

TCSETSF = 21_508

TCGETA = 21_509

TCSETA = 21_510

TCSETAW = 21_511

TCSETAF = 21_512

TCSBRK = 21_513

TCXONC = 21_514

TCFLSH = 21_515

TIOCEXCL = 21_516

TIOCNXCL = 21_517

TIOCSCTTY = 21_518

TIOCGPGRP = 21_519

TIOCSPGRP = 21_520

TIOCOUTQ = 21_521

TIOCSTI = 21_522

TIOCGWINSZ = 21_523

TIOCSWINSZ = 21_524

TIOCMGET = 21_525

TIOCMBIS = 21_526

TIOCMBIC = 21_527

TIOCMSET = 21_528

TIOCGSOFTCAR = 21_529

TIOCSSOFTCAR = 21_530

FIONREAD = 21_531

TIOCLINUX = 21_532

TIOCCONS = 21_533

TIOCGSERIAL = 21_534

TIOCSSERIAL = 21_535

TIOCPKT = 21_536

FIONBIO = 21_537

TIOCNOTTY = 21_538

TIOCSETD = 21_539

TIOCGETD = 21_540

TCSBRKP = 21_541

TIOCSBRK = 21_543

TIOCCBRK = 21_544

TIOCGSID = 21_545

TIOCGRS485 = 21_550

TIOCSRS485 = 21_551

TCGETX = 21_554

TCSETX = 21_555

TCSETXF = 21_556

TCSETXW = 21_557

TIOCVHANGUP = 21_559

FIONCLEX = 21_584

FIOCLEX = 21_585

FIOASYNC = 21_586

TIOCSERCONFIG = 21_587

TIOCSERGWILD = 21_588

TIOCSERSWILD = 21_589

TIOCGLCKTRMIOS = 21_590

TIOCSLCKTRMIOS = 21_591

TIOCSERGSTRUCT = 21_592

TIOCSERGETLSR = 21_593

TIOCSERGETMULTI = 21_594

TIOCSERSETMULTI = 21_595

TIOCMIWAIT = 21_596

TIOCGICOUNT = 21_597

FIOQSIZE = 21_600

FIOSETOWN = 35_073

SIOCSPGRP = 35_074

FIOGETOWN = 35_075

SIOCGPGRP = 35_076

SIOCATMARK = 35_077

SOCK_IOC_TYPE = 137

EPOLL_IOC_TYPE = 138

SIOCGSTAMP = 35_078

SIOCGSTAMPNS = 35_079

SIOCADDRT = 35_083

SIOCDELRT = 35_084

SIOCRTMSG = 35_085

SIOCGIFNAME = 35_088

SIOCSIFLINK = 35_089

SIOCGIFCONF = 35_090

SIOCGIFFLAGS = 35_091

SIOCSIFFLAGS = 35_092

SIOCGIFADDR = 35_093

SIOCSIFADDR = 35_094

SIOCGIFDSTADDR = 35_095

SIOCSIFDSTADDR = 35_096

SIOCGIFBRDADDR = 35_097

SIOCSIFBRDADDR = 35_098

SIOCGIFNETMASK = 35_099

SIOCSIFNETMASK = 35_100

SIOCGIFMETRIC = 35_101

SIOCSIFMETRIC = 35_102

SIOCGIFMEM = 35_103

SIOCSIFMEM = 35_104

SIOCGIFMTU = 35_105

SIOCSIFMTU = 35_106

SIOCSIFNAME = 35_107

SIOCSIFHWADDR = 35_108

SIOCGIFENCAP = 35_109

SIOCSIFENCAP = 35_110

SIOCGIFHWADDR = 35_111

SIOCGIFSLAVE = 35_113

SIOCSIFSLAVE = 35_120

SIOCADDMULTI = 35_121

SIOCDELMULTI = 35_122

SIOCGIFINDEX = 35_123

SIOCSIFPFLAGS = 35_124

SIOCGIFPFLAGS = 35_125

SIOCDIFADDR = 35_126

SIOCSIFHWBROADCAST = 35_127

SIOCGIFCOUNT = 35_128

SIOCGIFBR = 35_136

SIOCSIFBR = 35_137

SIOCGIFTXQLEN = 35_138

SIOCSIFTXQLEN = 35_139

SIOCETHTOOL = 35_142

SIOCGMIIPHY = 35_143

SIOCGMIIREG = 35_144

SIOCSMIIREG = 35_145

SIOCWANDEV = 35_146

SIOCOUTQNSD = 35_147

SIOCGSKNS = 35_148

SIOCDARP = 35_155

SIOCGARP = 35_156

SIOCSARP = 35_157

SIOCDRARP = 35_168

SIOCGRARP = 35_169

SIOCSRARP = 35_170

SIOCGIFMAP = 35_184

SIOCSIFMAP = 35_185

SIOCADDDLCI = 35_200

SIOCDELDLCI = 35_201

SIOCGIFVLAN = 35_202

SIOCSIFVLAN = 35_203

SIOCBONDENSLAVE = 35_216

SIOCBONDRELEASE = 35_217

SIOCBONDSETHWADDR = 35_218

SIOCBONDSLAVEINFOQUERY = 35_219

SIOCBONDINFOQUERY = 35_220

SIOCBONDCHANGEACTIVE = 35_221

SIOCBRADDBR = 35_232

SIOCBRDELBR = 35_233

SIOCBRADDIF = 35_234

SIOCBRDELIF = 35_235

SIOCSHWTSTAMP = 35_248

SIOCGHWTSTAMP = 35_249

SIOCDEVPRIVATE = 35_312

SIOCPROTOPRIVATE = 35_296

Implementations

impl IoctlRequest

pub const TIOCINQ: Self = _

pub const SIOGIFINDEX: Self = _