linux_api
Module mman

Structs

MRemapFlags

Flags used with mremap. u64 to match the x86-64 mremap syscall parameter: https://github.com/torvalds/linux/tree/v6.3/mm/mremap.c#L895

MapFlags

Map flags, as used with mmap. These are u64 to match the x86-64 mmap syscall parameter: https://github.com/torvalds/linux/tree/v6.3/arch/x86/kernel/sys_x86_64.c#L86

ProtFlags

Prot flags, as used with mmap. These are u64 to match the x86-64 mmap syscall parameter: https://github.com/torvalds/linux/tree/v6.3/arch/x86/kernel/sys_x86_64.c#L86

Functions

mmap⚠

Make the mmap syscall, with a thin layer of type-safety over mmap_raw. See mmap(2).

mmap_raw⚠

Make the mmap syscall. See mmap(2).

mprotect⚠

Make the mprotect syscall, with a thin layer of type-safety over mprotect_raw. See mprotect(2).

mprotect_raw⚠

Make the mprotect syscall. See mprotect(2).

munmap⚠

Make the munmap syscall, with a thin layer of type-safety over munmap_raw. See munmap(2).

munmap_raw⚠

Make the munmap syscall. See munmap(2).