linux_api

Module time

Structs

ClockNanosleepFlags

Valid flags passed to clock_nanosleep(2).

linux___kernel_old_itimerval

linux___kernel_old_timeval

linux___kernel_timespec

linux_itimerspec

linux_itimerval

linux_timespec

linux_timeval

Enums

ClockId

Clocks

ITimerId

Interval timers

Functions

alarm

Make an alarm syscall.

alarm_raw

Raw alarm syscall. Permits u64 arg and return value for generality with the general syscall ABI, but note that the alarm syscall definition itself uses u32.

clock_gettime

clock_gettime_raw

getitimer

Make a getitimer syscall.

getitimer_raw⚠

Make a getitimer syscall.

setitimer

Make a setitimer syscall.

setitimer_raw⚠

Make a setitimer syscall.

Type Aliases

itimerspec

itimerval

kernel_old_itimerval

kernel_old_timeval

kernel_timespec

linux___kernel_clockid_t

timespec

timeval