linux_api::time

Function setitimer_raw

pub unsafe fn setitimer_raw(
    which: c_int,
    new_value: *const kernel_old_itimerval,
    old_value: *mut kernel_old_itimerval,
) -> Result<(), Errno>

Make a setitimer syscall.

Safety

old_value must be safe for the kernel to write to, or NULL.

An invalid or inaccessible new_value isn’t a safety violation, but may cause the syscall to fail e.g. with EFAULT.